requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cobble==0.1.4
cryptography==45.0.5
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
google_search_results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
mammoth==1.9.1
markdown-it-py==3.0.0
markdownify==1.1.0
mdurl==0.1.2
numpy==2.3.1
openai==1.95.1
openpyxl==3.1.5
pandas==2.3.1
pathvalidate==3.3.1
pdfminer.six==20250506
pillow==11.3.0
puremagic==1.30
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5
youtube-transcript-api==1.1.1
